lib: fix multipart state transition in boundary parsing
When multipart boundary parsing in aborted due to an unexpected character,
we must transition back to either PART_DATA or BOUNDARY_START depending on
whether the currently processed character is a carriage return or not.

Before this fix, the parser state would go out of sync with the input data
when parsing a multiline value with the last line being a sole non-whitespace
character.

Fixes: #2
Signed-off-by: Jo-Philipp Wich <jo@mein.io>