lib: fix potentially lost bytes in boundary parsing across buffer limits
Only store non-carriage-return-chars terminating the boundary scan state in
the lookbehind buffer to push them out along with the other lookbehind bytes
in the subsequent callback invocation.

The previous code always stored the current char in the lookbehind buffer
and relied on the subsequent boundary-start or part-data states to handle
it, but the part-data state didn't consider the lookbehind buffer at all
which effectively resulted in lost bytes when a carriage return followed by
an arbitrary byte occured at the very end of a parse buffer.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>