multipart-parser: properly handle parts with trailing CR
A CR within the part data stream causes the parser to switch into the
boundary begin state, if the subsequent byte is not a newline it is
consumed and flushed as/appended to the part data. When a sequence such
as CR CR NL occurs in the data stream, the second CR is not treated
as potential boundary start, causing the parser to remain stuck in the
data read state until hitting EOF.

Solve this issue by only transitioning back to the part data state and
consuming the byte if it is not a CR and otherwise remaining in the
boundary start state and storing the CR in the lookbehind buffer.

Fixes: https://github.com/openwrt/luci/issues/6282
Signed-off-by: Jo-Philipp Wich <jo@mein.io>